README.md
Ideally prisoner would be able to move on with there lives after they have served their time. Unfortunately this is not the case because they are not given a chance at a new life. That is the goal of this project I want this to be an opportunity for ex felons to get put on a good path.