re-entry
Ideally prisoner would be able to go on and live their lives after leaving prison however this is not the case. Some try to change their lives around for the better but they are not given an opportunity to change themselves. This is where my project comes in I want to be the opportunity for them and help them grow and change.